Move browser module from experimental namespace
